requirements.txt
fastapi
uvicorn
python-multipart
pydantic
langchain
langchain-community
langchain-google-genai
sentence-transformers
faiss-cpu
pypdf
python-dotenv
slowapi